+++ service/aws/iam/iam_test.go
@@ -2,19 +2,33 @@ package iam
 
 import "github.com/stretchr/testify/mock"
 
+const (
+	testUserName = "test@example.com"
+	testAlias    = "aws-test-account-alias"
+	testARN      = "arn:aws:iam::000123456789:user/test@example.com"
+	testUserID   = "ABCDEFGHIJKLMNOPQRST"
+	testErrMsg   = "simulated error"
+)
+
+var (
+	testUsers = []string{"test1@example.com", "test2@example.com"}
+)
+
 /**
 * Mock using testify Framework
 * Technique : Interface Substitution
  */
 
-const testUserName = "test@example.com"
-const testAlias = "aws-test-account-alias"
-const testARN = "arn:aws:iam::000123456789:user/test@example.com"
-const testUserID = "ABCDEFGHIJKLMNOPQRST"
-
-var testUsers = []string{"test1@example.com", "test2@example.com"}
-
 // To mock AWS operations
 type MockClient struct {
 	mock.Mock
+}
+
+/**
+* Mock via manual creation - Just For Reference
+* Technique : Interface Substitution
+ */
+
+type MockReciever struct {
+	wantErr error
 }
